remove component.json
I will never update this file.  It's just going to be a source of
friction and confusion.  There's already a package manager for
JavaScript, and it's the one I wrote and built a company around.  I will
continue to keep package.json up to date, because that's how my workflow
works.

As the maintainer of this package, I get zero value out of having a
component.json file, and my inability to remember to keep it in sync
will only cause problems.

If anyone is upset by this, then they are more than welcome to maintain
a fork of sax-js that adds a component.json file, or submit a patch to
bower or component or whatever to have it read package.json for package
metadata.

Fixes #123